✨ Refacto de ExonérationCovid
